fix(web): move the chart to components folder
