+++ web/components/PoissonDistributionChart.jsx
